further refactorings and simplifications
