+++ internal/rules/config/parser.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package rule
+package config
 
 import (
 	"errors"
@@ -23,13 +23,12 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/dadrus/heimdall/internal/heimdall"
-	"github.com/dadrus/heimdall/internal/rules/config"
 	"github.com/dadrus/heimdall/internal/x/errorchain"
 )
 
 var ErrEmptyRuleSet = errors.New("empty rule set")
 
-func ParseRules(contentType string, reader io.Reader) (*config.RuleSet, error) {
+func ParseRules(contentType string, reader io.Reader) (*RuleSet, error) {
 	switch contentType {
 	case "application/json":
 		fallthrough
@@ -48,10 +47,10 @@ func ParseRules(contentType string, reader io.Reader) (*config.RuleSet, error) {
 	}
 }
 
-func parseYAML(reader io.Reader) (*config.RuleSet, error) {
+func parseYAML(reader io.Reader) (*RuleSet, error) {
 	var (
 		rawConfig map[string]any
-		ruleSet   config.RuleSet
+		ruleSet   RuleSet
 	)
 
 	dec := yaml.NewDecoder(reader)
@@ -63,7 +62,7 @@ func parseYAML(reader io.Reader) (*config.RuleSet, error) {
 		return nil, err
 	}
 
-	err := config.DecodeConfig(rawConfig, &ruleSet)
+	err := DecodeConfig(rawConfig, &ruleSet)
 
 	return &ruleSet, err
 }
