+++ internal/rules/config/parser_test.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package rule
+package config
 
 import (
 	"bytes"
@@ -24,7 +24,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/dadrus/heimdall/internal/heimdall"
-	"github.com/dadrus/heimdall/internal/rules/config"
 )
 
 func TestParseRules(t *testing.T) {
@@ -34,13 +33,13 @@ func TestParseRules(t *testing.T) {
 		uc          string
 		contentType string
 		content     []byte
-		assert      func(t *testing.T, err error, ruleSet *config.RuleSet)
+		assert      func(t *testing.T, err error, ruleSet *RuleSet)
 	}{
 		{
 			uc:          "unsupported content type and not empty contents",
 			contentType: "foobar",
 			content:     []byte(`foo: bar`),
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.Error(t, err)
@@ -51,7 +50,7 @@ func TestParseRules(t *testing.T) {
 		{
 			uc:          "unsupported content type and empty contents",
 			contentType: "foobar",
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.ErrorIs(t, err, ErrEmptyRuleSet)
@@ -66,7 +65,7 @@ func TestParseRules(t *testing.T) {
 "name": "foo",
 "rules": [{"id": "bar"}]
 }`),
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -79,7 +78,7 @@ func TestParseRules(t *testing.T) {
 		{
 			uc:          "JSON content type and empty contents",
 			contentType: "application/json",
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.ErrorIs(t, err, ErrEmptyRuleSet)
@@ -95,7 +94,7 @@ name: foo
 rules:
 - id: bar
 `),
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.NoError(t, err)
@@ -108,7 +107,7 @@ rules:
 		{
 			uc:          "YAML content and empty contents",
 			contentType: "application/yaml",
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.ErrorIs(t, err, ErrEmptyRuleSet)
@@ -132,11 +131,11 @@ func TestParseYAML(t *testing.T) {
 	for _, tc := range []struct {
 		uc     string
 		conf   []byte
-		assert func(t *testing.T, err error, ruleSet *config.RuleSet)
+		assert func(t *testing.T, err error, ruleSet *RuleSet)
 	}{
 		{
 			uc: "empty rule set spec",
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.ErrorIs(t, err, ErrEmptyRuleSet)
@@ -145,7 +144,7 @@ func TestParseYAML(t *testing.T) {
 		{
 			uc:   "invalid rule set spec",
 			conf: []byte(`foo: bar`),
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.Error(t, err)
@@ -159,7 +158,7 @@ name: foo
 rules:
 - id: bar
 `),
-			assert: func(t *testing.T, err error, ruleSet *config.RuleSet) {
+			assert: func(t *testing.T, err error, ruleSet *RuleSet) {
 				t.Helper()
 
 				require.NoError(t, err)
